pyproject.toml
[tool.poetry]
name = "api-it-solutions"
version = "0.1.0"
description = ""
authors = ["nikita-r <gfgfgf354@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.3"
drf-spectacular = "^0.27.2"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
